streamlined the Credentials interface; ensure support for remote credentials
