+++ basculehash/hasher.go
@@ -45,50 +45,3 @@ var defaultHash HasherComparer = Bcrypt{}
 func Default() HasherComparer {
 	return defaultHash
 }
-
-// Matches tests if the given Comparer strategy indicates a match
-// between the plaintext and the digest.  If cmp is nil, the
-// Default() is used.
-func Matches(cmp Comparer, plaintext []byte, d Digest) error {
-	if cmp == nil {
-		cmp = Default()
-	}
-
-	return cmp.Matches(plaintext, d)
-}
-
-// Matcher is anything that can test if a principal's digest matches a digest.
-type Matcher interface {
-	// Matches checks the associated digest with the given plaintext.
-	Matches(cmp Comparer, principal string, plaintext []byte) error
-}
-
-// Credentials is a set of principals and associated digests. Each principal may
-// have exactly zero (0) or one (1) associated Digest.
-//
-// All implementations of this interface in this package can be marshaled and
-// unmarshaled from JSON.
-type Credentials interface {
-	Matcher
-
-	// Len returns the count of principals in this set.
-	Len() int
-
-	// Get returns the Digest associated with the given Principal.
-	// This method returns false if the principal did not exist.
-	Get(principal string) (d Digest, exists bool)
-
-	// Set associates a principal with a Digest.  If the principal already
-	// exists, its digest is replaced.
-	//
-	// This method does not retain d.  An copy is made and stored internally.
-	Set(principal string, d Digest)
-
-	// Delete removes the principal from this set.  If the principal did
-	// not exist, it returns false.
-	Delete(principal string) (d Digest, existed bool)
-
-	// Update performs a bulk update of these credentials. Any existing
-	// principals are replaced.
-	Update(p Principals)
-}
