+++ basculehash/principals.go
@@ -3,7 +3,9 @@
 
 package basculehash
 
-import "fmt"
+import (
+	"context"
+)
 
 // Principals is a Credentials implementation that is a simple map
 // of principals to digests.  This type is not safe for concurrent
@@ -13,54 +15,31 @@ import "fmt"
 // or protected from concurrent updates by some other means.
 type Principals map[string]Digest
 
-var _ Matcher = Principals{}
 var _ Credentials = Principals{}
-
-// Len returns the number of principals in this set.
-func (p Principals) Len() int {
-	return len(p)
-}
 
 // Get returns the Digest associated with the principal.  This method
 // returns false if the principal did not exist.
-func (p Principals) Get(principal string) (d Digest, exists bool) {
+func (p Principals) Get(_ context.Context, principal string) (d Digest, exists bool) {
 	d, exists = p[principal]
 	return
 }
 
 // Set adds or replaces the given principal and its associated digest.
-func (p Principals) Set(principal string, d Digest) {
+func (p Principals) Set(_ context.Context, principal string, d Digest) {
 	p[principal] = d.Copy()
 }
 
-// Delete removes the given principal from this set, returning any existing
-// Digest and an indicator of whether it existed.
-func (p Principals) Delete(principal string) (d Digest, existed bool) {
-	if d, existed = (p)[principal]; existed {
-		delete(p, principal)
+// Delete removes the given principal(s) from this set.
+func (p Principals) Delete(_ context.Context, principals ...string) {
+	for _, toDelete := range principals {
+		delete(p, toDelete)
 	}
-
-	return
 }
 
 // Update performs a bulk update of credentials. Each digest is copied
 // before storing in this instance.
-func (p Principals) Update(more Principals) {
+func (p Principals) Update(_ context.Context, more Principals) {
 	for principal, digest := range more {
 		p[principal] = digest.Copy()
 	}
-}
-
-// Matches tests if a given principal's password matches the associated
-// digest.  If no such principal exists, this method returns bascule.ErrBadCredentials.
-//
-// If cmp is nil, DefaultComparer is used.
-func (p Principals) Matches(cmp Comparer, principal string, plaintext []byte) (err error) {
-	if d, exists := p[principal]; exists {
-		err = Matches(cmp, plaintext, d)
-	} else {
-		err = fmt.Errorf("No such principal: %s", principal)
-	}
-
-	return
 }
